Minor Code Fix

+++ Engine/common_dx/source/DXMipmapGenerator.cpp
@@ -121,7 +121,7 @@ void DXMipmapGenerator::Generate(const ComPtr<ID3D12Device>& device, const ComPt
         cpuHandle.Offset(1, m_descriptorSize);
         gpuHandle.Offset(1, m_descriptorSize);
 
-		float texelSize[2] = { 1.0f / dstWidth, 1.0f / dstHeight };
+		float texelSize[2] = { 1.0f / static_cast<float>(dstWidth), 1.0f / static_cast<float>(dstHeight) };
         commandList->SetComputeRoot32BitConstants(0, 2, texelSize, 0);
 
 		UINT dispatchX = (dstWidth + 7) / 8;
